perf: use less promises for ReadableStream
