+++ package.json
@@ -116,6 +116,7 @@
     "abort-controller": "^3.0.0",
     "borp": "^0.20.0",
     "c8": "^10.0.0",
+    "count-promises": "^1.0.1",
     "cross-env": "^10.0.0",
     "dns-packet": "^5.4.0",
     "esbuild": "^0.25.2",
